refactor(job card): refactor of modal base structure
- refactor of base modal structure to ensure more flexibility and improve reusability.
