+++ src/components/JobCard/ModalHolder/ModalHolder.module.scss
@@ -1,0 +1,5 @@
+.contentContainer {
+  background-color: white;
+  width: 100vw;
+  height: 80vh;
+}
