+++ src/components/Modal/Modal.module.scss
@@ -1,0 +1,11 @@
+@import "./../../styles/scss/index.scss";
+.container {
+  position: fixed;
+  top: 0;
+  left: 0;
+  right: 0;
+  bottom: 0;
+  background-color: rgba(0, 0, 0, 0.7);
+  z-index: 1000;
+  @include displayFlex($align: center);
+}
