+++ src/components/index.js
@@ -10,4 +10,5 @@ export { default as DashCard } from "./DashCard/DashCard";
 export { default as FormProvider } from "./FormProvider/FormProvider";
 export { default as FormClient } from "./FormClient/FormClient";
 export { default as FormSignupSuccessful } from "./FormSignupSuccessful/FormSignupSuccessful";
-export { default as CardModal } from "./CardModal/CardModal";
+export { default as Modal } from "./Modal/Modal";
+export { default as ModalHolder } from "./JobCard/ModalHolder/ModalHolder";
